64bit warning fix
git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/loggertools/trunk@110 fae35197-53e2-0310-b5ce-cc51903860e9

+++ src/fwd.c
@@ -79,7 +79,7 @@ static void dump_char(struct dump *d, const char *prefix,
     }
 
     if (d->column == 0)
-        printf("%s 0x%04x: ", prefix, d->offset);
+        printf("%s 0x%04x: ", prefix, (unsigned)d->offset);
 
     printf(" %02x", ch);
     fflush(stdout);
